Split MainViewTest to two files
