+++ app/src/test/kotlin/com/jereksel/libresubstratum/views/MainViewTest.kt
@@ -162,114 +162,6 @@ class MainViewTest: BaseRobolectricTest() {
 
         assertTrue(dialog.isShowing)
     }
-
-    @Test
-    fun `Lock should be visible if theme key cannot be found`() {
-
-        RxJavaPlugins.setComputationSchedulerHandler { Schedulers.trampoline() }
-        RxJavaPlugins.setIoSchedulerHandler { Schedulers.trampoline() }
-
-        val encryptedApp = InstalledTheme("id1", "name1", "author1", false, "2", FutureTask { null })
-        whenever(presenter.getKeyPair("id1")).thenReturn(null)
-        activity.addApplications(listOf(encryptedApp))
-
-        recyclerView.measure(0, 0);
-        recyclerView.layout(0, 0, 100, 10000);
-
-        val view = recyclerView.getChildAt(0).find<ImageView>(R.id.lock)
-
-        assertEquals(view.visibility, View.VISIBLE)
-    }
-
-    @Test
-    fun `Lock should not be visible if theme is not encrypted`() {
-
-        RxJavaPlugins.setComputationSchedulerHandler { Schedulers.trampoline() }
-        RxJavaPlugins.setIoSchedulerHandler { Schedulers.trampoline() }
-
-        val encryptedApp = InstalledTheme("id1", "name1", "author1", false, "2", FutureTask { null })
-        whenever(presenter.getKeyPair("id1")).thenReturn(KeyPair.EMPTYKEY)
-        activity.addApplications(listOf(encryptedApp))
-
-        recyclerView.measure(0, 0);
-        recyclerView.layout(0, 0, 100, 10000);
-
-        val view = recyclerView.getChildAt(0).find<ImageView>(R.id.lock)
-
-        assertEquals(view.visibility, View.GONE)
-    }
-
-    @Test
-    fun `When clicking on lock toast is shown`() {
-
-        RxJavaPlugins.setComputationSchedulerHandler { Schedulers.trampoline() }
-        RxJavaPlugins.setIoSchedulerHandler { Schedulers.trampoline() }
-
-        val encryptedApp = InstalledTheme("id1", "name1", "author1", false, "2", FutureTask { null })
-        activity.addApplications(listOf(encryptedApp))
-
-        recyclerView.measure(0, 0);
-        recyclerView.layout(0, 0, 100, 10000);
-
-        recyclerView.getChildAt(0).find<ImageView>(R.id.lock).performClick()
-
-        assertEquals(RuntimeEnvironment.application.getString(R.string.unsupported_template_toast), ShadowToast.getTextOfLatestToast())
-
-    }
-
-    @Test
-    fun `When waiting for keypair spinner is shown`() {
-
-        val testScheduler = TestScheduler()
-
-        RxJavaPlugins.reset()
-        RxJavaPlugins.setComputationSchedulerHandler { testScheduler }
-        RxJavaPlugins.setIoSchedulerHandler { testScheduler }
-
-        val encryptedApp = InstalledTheme("id1", "name1", "author1", false, "2", FutureTask { null })
-        whenever(presenter.getKeyPair("id1")).thenReturn(KeyPair.EMPTYKEY)
-        activity.addApplications(listOf(encryptedApp))
-
-        recyclerView.measure(0, 0);
-        recyclerView.layout(0, 0, 100, 10000);
-
-        val view = recyclerView.getChildAt(0).find<ProgressBar>(R.id.progressBar)
-
-        assertEquals(view.visibility, View.VISIBLE)
-    }
-
-    @Test
-    fun `When keypair exists spinner is hidden`() {
-
-        RxJavaPlugins.setComputationSchedulerHandler { Schedulers.trampoline() }
-        RxJavaPlugins.setIoSchedulerHandler { Schedulers.trampoline() }
-
-        val encryptedApp = InstalledTheme("id1", "name1", "author1", false, "2", FutureTask { null })
-        whenever(presenter.getKeyPair("id1")).thenReturn(KeyPair.EMPTYKEY)
-        activity.addApplications(listOf(encryptedApp))
-
-        recyclerView.measure(0, 0);
-        recyclerView.layout(0, 0, 100, 10000);
-
-        val view = recyclerView.getChildAt(0).find<ProgressBar>(R.id.progressBar)
-
-        assertEquals(view.visibility, View.GONE)
-    }
-
-    @Test
-    fun `dismissDialog() should dismiss any dialog`() {
-
-        val message = "Very important message"
-
-        activity.showUndismissableDialog(message)
-
-        assertTrue(ShadowDialog.getLatestDialog().isShowing)
-
-        activity.dismissDialog()
-
-        assertFalse(ShadowDialog.getLatestDialog().isShowing)
-    }
-
 }
